subscription: refactoring to use the new structure
