+++ azurerm/internal/provider/services.go
@@ -49,6 +49,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/redis"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/relay"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/resource"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/subscription"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/trafficmanager"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/web"
 )
@@ -103,6 +104,7 @@ func SupportedServices() []common.ServiceRegistration {
 		relay.Registration{},
 		resource.Registration{},
 
+		subscription.Registration{},
 		trafficmanager.Registration{},
 		web.Registration{},
 	}
