+++ azurerm/internal/services/subscription/client/client.go
@@ -1,4 +1,4 @@
-package subscription
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2018-06-01/subscriptions"
@@ -9,7 +9,7 @@ type Client struct {
 	Client *subscriptions.Client
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	client := subscriptions.NewClientWithBaseURI(o.ResourceManagerEndpoint)
 	o.ConfigureClient(&client.Client, o.ResourceManagerAuthorizer)
